requirements.txt
streamlit
tweepy
vaderSentiment
tweet-preprocessor
matplotlib